requirement.txt
openai-whisper
moviepy
langchain_openai
python-dotenv